requirements.txt
python-telegram-bot==21.6
py-gemini-watermark-remover==0.2.9
opencv-python-headless
numpy
